090121-03:00
Expanded the on_receive() function to allow the client to detect when the server is telling them they have fired a bullet or quit.

This functionality enabled sound effects to be added to the game. Further testing is of course required to see if this all works as expected across a network between two clients.

Set the IPv4 address of the client to match that of the server's.